% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{ffcbs_api}
\alias{ffcbs_api}
\title{CBS Sports Fantasy Football API}
\usage{
ffcbs_api(path, query = NULL)
}
\arguments{
\item{path}{path to pull}

\item{query}{query parameters}
}
\description{
CBS Sports Fantasy Football API
}
