% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.r
\name{ffcbs_projections}
\alias{ffcbs_projections}
\title{Projections}
\usage{
ffcbs_projections(season, week, pos = c("QB", "RB", "WR", "TE", "K",
  "DST"))
}
\arguments{
\item{season}{integer year}

\item{week}{integer week (0 - 17)}

\item{pos}{character position}
}
\description{
Projections
}
