% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.r
\name{tidy_projections}
\alias{tidy_projections}
\title{Tidy projections}
\usage{
tidy_projections(x, pos)
}
\arguments{
\item{x}{data frame of projections}

\item{pos}{position}
}
\description{
Tidy projections
}
\keyword{internal}
